Add initail html for activity log item component
Refactor into its own folders

Updates Activity log item styles
