+++ client/my-sites/stats/activity-log-item/style.scss
