--- client/my-sites/stats/activity-log/item.jsx
@@ -1,20 +1,0 @@
-/**
- * External dependencies
- */
-import React from 'react';
-import { localize } from 'i18n-calypso';
-/**
- * Internal dependencies
- */
-import Card from 'components/card/compact';
-
-const ActivityLogItem = React.createClass( {
-
-	render() {
-		return (
-			<Card>hello there</Card>
-		);
-	}
-} );
-
-export default localize( ActivityLogItem );
